scale_bits

Module bits

This module exposes a Bits type, which is a small, simple bit store which is SCALE compatible with BitVec<u8, Lsb0>.

Structs

Bits

This represents a sequence of boolean values, packed into bits.

BitsIntoIter

Returned from calling into_iter on Bits via the [std::iter::IntoIterator] trait. Allows iteration over each stored bit.

BitsIter

Returned from calling Bits::iter(). Allows iteration over each stored bit.